requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.2
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
optuna==4.0.0
python-multipart==0.0.17
jinja2==3.1.4
openpyxl==3.1.5
python-dotenv==1.0.1
replicate==0.32.1
requests==2.32.3
